Chore/qol adder cost (#182)
* QoL improvements to UX, reverted to Long for custom adder amounts in back-end

* chore: price table custom value showing, fixed v-chip display

+++ src/main/java/com/albatross/api/v1/company/blueraven/models/ProposalAdderRequest.java
@@ -15,7 +15,7 @@ public class ProposalAdderRequest {
   @NoArgsConstructor
   public static class AdderItem {
     private Long id;
-    private BigDecimal customAdderAmount;
+    private Long customAdderAmount;
     private String adderType;
     private String fieldName;
     private Boolean applied = true;
